requirements.txt
matplotlib==2.2.3
numpy==1.18.5
pickleshare==0.7.5
PyYAML==5.3.1
torch==1.5.1
torchvision==0.6.1
